Add AbstractBeaconStateImpl base class for common logic
