+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/state/beaconstate/AbstractBeaconStateImpl.java
@@ -1,0 +1,87 @@
+/*
+ * Copyright 2019 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.spec.datastructures.state.beaconstate;
+
+import com.google.common.base.MoreObjects.ToStringHelper;
+import tech.pegasys.teku.ssz.backing.SszContainer;
+import tech.pegasys.teku.ssz.backing.SszData;
+import tech.pegasys.teku.ssz.backing.cache.IntCache;
+import tech.pegasys.teku.ssz.backing.cache.SoftRefIntCache;
+import tech.pegasys.teku.ssz.backing.schema.AbstractSszContainerSchema;
+import tech.pegasys.teku.ssz.backing.schema.SszCompositeSchema;
+import tech.pegasys.teku.ssz.backing.tree.TreeNode;
+import tech.pegasys.teku.ssz.backing.view.SszContainerImpl;
+
+public abstract class AbstractBeaconStateImpl<TMutable extends MutableBeaconState>
+    extends SszContainerImpl implements BeaconState, BeaconStateCache {
+
+  private final TransitionCaches transitionCaches;
+
+  protected AbstractBeaconStateImpl(final BeaconStateSchema<?, ?> schema) {
+    super(schema);
+    transitionCaches = TransitionCaches.createNewEmpty();
+  }
+
+  protected AbstractBeaconStateImpl(
+      SszCompositeSchema<?> type,
+      TreeNode backingNode,
+      IntCache<SszData> cache,
+      TransitionCaches transitionCaches) {
+    super(type, backingNode, cache);
+    this.transitionCaches = transitionCaches;
+  }
+
+  protected AbstractBeaconStateImpl(
+      AbstractSszContainerSchema<? extends SszContainer> type, TreeNode backingNode) {
+    super(type, backingNode);
+    transitionCaches = TransitionCaches.createNewEmpty();
+  }
+
+  @Override
+  public <E1 extends Exception, E2 extends Exception, E3 extends Exception> BeaconState updated(
+      Mutator<MutableBeaconState, E1, E2, E3> mutator) throws E1, E2, E3 {
+    MutableBeaconState writableCopy = createWritableCopyPriv();
+    mutator.mutate(writableCopy);
+    return writableCopy.commitChanges();
+  }
+
+  @Override
+  public int hashCode() {
+    return BeaconStateInvariants.hashCode(this);
+  }
+
+  @Override
+  public boolean equals(Object obj) {
+    return BeaconStateInvariants.equals(this, obj);
+  }
+
+  @Override
+  public TransitionCaches getTransitionCaches() {
+    return transitionCaches;
+  }
+
+  @Override
+  protected IntCache<SszData> createCache() {
+    return new SoftRefIntCache<>(super::createCache);
+  }
+
+  @Override
+  public String toString() {
+    return BeaconStateInvariants.toString(this, this::describeCustomFields);
+  }
+
+  protected abstract void describeCustomFields(ToStringHelper stringBuilder);
+
+  protected abstract TMutable createWritableCopyPriv();
+}
