+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/state/beaconstate/BeaconStateInvariants.java
@@ -15,7 +15,11 @@ package tech.pegasys.teku.spec.datastructures.state.beaconstate;
 
 import static com.google.common.base.Preconditions.checkState;
 
+import com.google.common.base.MoreObjects;
+import com.google.common.base.MoreObjects.ToStringHelper;
 import java.util.List;
+import java.util.Objects;
+import java.util.function.Consumer;
 import org.apache.tuweni.bytes.Bytes;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.ssz.backing.schema.SszPrimitiveSchemas;
@@ -66,5 +70,53 @@ public class BeaconStateInvariants {
     // Extract slot data
     final Bytes slotData = bytes.slice(offset, size);
     return SLOT_SCHEMA.sszDeserialize(slotData).get();
+  }
+
+  static boolean equals(BeaconState state, Object obj) {
+    if (Objects.isNull(obj)) {
+      return false;
+    }
+
+    if (state == obj) {
+      return true;
+    }
+
+    if (!(obj instanceof BeaconState)) {
+      return false;
+    }
+
+    BeaconState other = (BeaconState) obj;
+    return state.hashTreeRoot().equals(other.hashTreeRoot());
+  }
+
+  static int hashCode(BeaconState state) {
+    return state.hashTreeRoot().slice(0, 4).toInt();
+  }
+
+  static String toString(BeaconState state, final Consumer<ToStringHelper> modifier) {
+    final ToStringHelper builder =
+        MoreObjects.toStringHelper(state)
+            .add("genesis_time", state.getGenesis_time())
+            .add("genesis_validators_root", state.getGenesis_validators_root())
+            .add("slot", state.getSlot())
+            .add("fork", state.getFork())
+            .add("latest_block_header", state.getLatest_block_header())
+            .add("block_roots", state.getBlock_roots())
+            .add("state_roots", state.getState_roots())
+            .add("historical_roots", state.getHistorical_roots())
+            .add("eth1_data", state.getEth1_data())
+            .add("eth1_data_votes", state.getEth1_data_votes())
+            .add("eth1_deposit_index", state.getEth1_deposit_index())
+            .add("validators", state.getValidators())
+            .add("balances", state.getBalances())
+            .add("randao_mixes", state.getRandao_mixes())
+            .add("slashings", state.getSlashings())
+            .add("justification_bits", state.getJustification_bits())
+            .add("previous_justified_checkpoint", state.getPrevious_justified_checkpoint())
+            .add("current_justified_checkpoint", state.getCurrent_justified_checkpoint())
+            .add("finalized_checkpoint", state.getFinalized_checkpoint());
+
+    modifier.accept(builder);
+    return builder.toString();
   }
 }
